Add doom example.
