+++ src/components/ProgramLoader/ProgramFileUpload.tsx
@@ -1,17 +1,14 @@
 import { useDropzone } from "react-dropzone";
-import { ProgramUploadFileInput, ProgramUploadFileOutput } from "./types";
-import { mapUploadFileInputToOutput } from "./utils";
-import { bytes, pvm } from "@typeberry/lib";
-import { ExpectedState, MemoryChunkItem, PageMapItem, RegistersArray } from "@/types/pvm.ts";
-import { SafeParseReturnType, z } from "zod";
+import { ProgramUploadFileOutput } from "./types";
+import { bytes } from "@typeberry/lib";
 import { useAppSelector } from "@/store/hooks";
 import { selectInitialState } from "@/store/debugger/debuggerSlice";
-import { cn, getAsChunks, getAsPageMap } from "@/lib/utils.ts";
+import { cn } from "@/lib/utils.ts";
 import { UploadCloud } from "lucide-react";
 import { Button } from "../ui/button";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { Input } from "../ui/input";
-import { decodeSpiWithMetadata } from "@/utils/spi";
+import { loadFileFromUint8Array } from "./loadingUtils";
 
 type ProgramFileUploadProps = {
   onFileUpload: (val: ProgramUploadFileOutput) => void;
@@ -19,148 +16,6 @@ type ProgramFileUploadProps = {
   isError: boolean;
   close?: () => void;
 };
-
-type RawProgramUploadFileInput = unknown;
-type ValidationResult = SafeParseReturnType<RawProgramUploadFileInput, ProgramUploadFileInput>;
-
-const validateJsonTestCaseSchema = (json: RawProgramUploadFileInput): ValidationResult => {
-  const pageMapSchema = z.object({
-    address: z.number(),
-    length: z.number(),
-    "is-writable": z.boolean(),
-  });
-
-  const memorySchema = z.object({
-    address: z.number(),
-    contents: z.array(z.number()),
-  });
-
-  const schema = z.object({
-    name: z.string(),
-    "initial-regs": z.array(z.number()).length(13),
-    "initial-pc": z.number(),
-    "initial-page-map": z.array(pageMapSchema),
-    "initial-memory": z.array(memorySchema),
-    "initial-gas": z.number(),
-    program: z.array(z.number()),
-    "expected-status": z.string(),
-    "expected-regs": z.array(z.number()),
-    "expected-pc": z.number(),
-    "expected-memory": z.array(memorySchema),
-    "expected-gas": z.number(),
-  });
-
-  return schema.safeParse(json);
-};
-
-const generateErrorMessageFromZodValidation = (result: ValidationResult): string => {
-  if (!result.error) return "Unknown error occurred";
-
-  const formattedErrors = result.error.errors.map((err) => {
-    const path = err.path.join(" > ") || "root";
-    return `Field: "${path}" - ${err.message}`;
-  });
-
-  return `File validation failed with the following errors:\n\n${formattedErrors.join("\n")}`;
-};
-
-function loadFileFromUint8Array(
-  loadedFileName: string,
-  uint8Array: Uint8Array,
-  spiArgsAsBytes: Uint8Array,
-  setError: (e?: string) => void,
-  onFileUpload: (d: ProgramUploadFileOutput) => void,
-  initialState: ExpectedState,
-) {
-  // reset error state
-  setError(undefined);
-
-  // Try to parse file as a JSON first
-  let jsonFile = null;
-  let stringContent = "";
-  let rawBytes = uint8Array;
-
-  try {
-    stringContent = new TextDecoder("utf-8").decode(uint8Array);
-  } catch (e) {
-    console.warn("not a string", e);
-  }
-
-  if (stringContent.startsWith("0x")) {
-    try {
-      rawBytes = bytes.BytesBlob.parseBlob(stringContent).raw;
-    } catch (e) {
-      console.warn("not a bytes blob", e);
-    }
-  } else {
-    try {
-      jsonFile = JSON.parse(stringContent);
-    } catch (e) {
-      console.warn("not a JSON", e);
-    }
-  }
-
-  if (jsonFile !== null) {
-    const result = validateJsonTestCaseSchema(jsonFile);
-    if (!result.success) {
-      console.warn("not a valid JSON", result.error);
-      const errorMessage = generateErrorMessageFromZodValidation(result);
-      setError(errorMessage || "");
-      return;
-    }
-
-    onFileUpload(mapUploadFileInputToOutput(jsonFile, "JSON test"));
-    return;
-  }
-
-  // Try to decode the program as an SPI
-  let spi = null;
-  try {
-    spi = decodeSpiWithMetadata(rawBytes, spiArgsAsBytes);
-  } catch (e) {
-    console.warn("not an SPI blob", e);
-  }
-  if (spi !== null) {
-    const { code, memory, registers } = spi;
-    const pageMap: PageMapItem[] = getAsPageMap(memory);
-    const chunks: MemoryChunkItem[] = getAsChunks(memory);
-
-    onFileUpload({
-      program: Array.from(code),
-      name: `${loadedFileName} [spi]`,
-      isSpi: true,
-      kind: "JAM SPI",
-      initial: {
-        regs: Array.from(registers).map((x) => BigInt(x as number | bigint)) as RegistersArray,
-        pc: 0,
-        pageMap,
-        memory: chunks,
-        gas: 10000n,
-      },
-    });
-    return;
-  }
-
-  // Finally try to load program as a Generic
-  let program = null;
-  try {
-    program = new pvm.ProgramDecoder(rawBytes);
-  } catch (e) {
-    console.warn("not a generic PVM", e);
-  }
-
-  if (program !== null) {
-    onFileUpload({
-      program: Array.from(rawBytes),
-      name: `${loadedFileName} [generic]`,
-      isSpi: false,
-      kind: "Generic PVM",
-      initial: initialState,
-    });
-  } else {
-    setError("Unrecognized program format (see console).");
-  }
-}
 
 export const ProgramFileUpload: React.FC<ProgramFileUploadProps> = ({ isError, onFileUpload, close, setError }) => {
   const initialState = useAppSelector(selectInitialState);
